refactor(watchlist): move under own folder
