+++ projects/client/src/routes/watchlist/+page.svelte
@@ -4,7 +4,7 @@
   import CoverImageSetter from "$lib/components/background/CoverImageSetter.svelte";
   import TraktPage from "$lib/components/layout/TraktPage.svelte";
   import { useUser } from "$lib/features/auth/stores/useUser";
-  import WatchlistList from "$lib/sections/lists/WatchlistList.svelte";
+  import WatchlistList from "$lib/sections/lists/watchlist/WatchlistList.svelte";
   import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
 
   const { current } = useUser();
